pyproject.toml
[tool.poetry]
name = "stunk"
version = "0.1.0"
description = ""
authors = ["wiredmatt <29718978+wiredmatt@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
yfinance = "^0.2.50"
pandas = "^2.2.3"
matplotlib = "^3.8.3"
seaborn = "^0.13.2"
newsapi-python = "^0.2.7"
python-dotenv = "^1.0.1"
python-telegram-bot = "^20.8"
sqlalchemy = "^2.0.28"
redis = "^5.0.3"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"

[tool.poetry.scripts]
init_db = "stunk.storage.init_db:main"
market_trend = "stunk.market_trend:main"
telegram_bot = "stunk.telegram_bot:run_bot"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
